Fixed some warnings: size_t/uint conversions and some deprecated method.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@24656 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/skeleton/object/skeleton.h
@@ -711,7 +711,7 @@ public:
   bool Initialize (iObjectRegistry* object_reg);
   bool HandleEvent (iEvent& ev);
 
-  virtual int GetFactoriesCount() { return factories.Length(); }
+  virtual size_t GetFactoriesCount() { return factories.Length(); }
   virtual iSkeletonFactory *CreateFactory(const char *name);
   virtual iSkeletonFactory *LoadFactory(const char * /*file_name*/) { return 0; }
   virtual iSkeletonFactory *FindFactory(const char * /*name*/) { return 0; }
